Added async loading of scenes. This doesn't work in web exports tho., because they are single threaded.
It is possible to export with multiple threads enabled. Itch io need a checkbox for this to be checked too.
(Something with 'Showdowbuffer).
However this results in gray screens if scenes change, and also on mac and linux the game might be buggy according to the documentation.
So for now, single threaded is the way for web exports.